Source for ArrayList2022 lectures and videos.
This is the source code used for the ArrayList playlist. It goes over the use of the Java ArrayList type using String, Wrapper classes, and custom classes. A Popup class is included for GUI implementation. Images of code are in the src/list/view/images directory and can be used for questions about the concepts.